+++ source/navigator/TaskNavigator.js
@@ -1,19 +1,13 @@
 import { createStackNavigator } from 'react-navigation-stack';
-import { TaskDetailScreen, TaskScreen } from 'app/screens';
+import { CaseArchiveScreen } from 'app/screens';
 
-export const TaskStack = {
-  Task: {
-    screen: TaskScreen,
-    navigationOptions: {
-      headerShown: false,
-    },
-  },
-  TaskDetails: {
-    screen: TaskDetailScreen,
+export const CaseStack = {
+  CaseArchive: {
+    screen: CaseArchiveScreen,
     navigationOptions: {
       headerShown: false,
     },
   },
 };
 
-export default createStackNavigator(TaskStack);
+export default createStackNavigator(CaseStack);
