+++ source/screens/index.js
@@ -3,6 +3,6 @@ export { default as LoginScreen } from './LoginScreen';
 export { default as ProfileScreen } from './ProfileScreen';
 export { default as SplashScreen } from './SplashScreen';
 export { default as TaskDetailScreen } from './TaskDetailScreen';
-export { default as TaskScreen } from './TaskScreen';
+export { default as CaseArchiveScreen } from './CaseArchiveScreen';
 export { default as HomeScreen } from './HomeScreen';
 export { default as FormCaseScreen } from './FormCaseScreen';
